chore: move annotated map to separate component
