+++ frontend/public/datasets.json
@@ -1,4 +1,9 @@
 [
   "datasets/raw-images/1_Sample_Data_Melamchi_Bazar_Area_Flooding",
-  "datasets/raw-images/4_Sample_Data_Jajarkot_Rukum_West/4202-337ac79e-92fc-4126-87e2-2f31298524bf_raw_images"
+  "datasets/raw-images/2_Sample_Data_Melamchi_Bazar_Access_Road_Section_Flooding",
+  "datasets/raw-images/3_Sample_Data_KTM_Flooding",
+  "datasets/raw-images/4_Sample_Data_Jajarkot_Rukum_West/4202-337ac79e-92fc-4126-87e2-2f31298524bf_raw_images",
+  "datasets/raw-images/4_Sample_Data_Jajarkot_Rukum_West/4205-1809d422-0af9-4a60-a1d8-7a3017baab98_raw_images",
+  "datasets/raw-images/4_Sample_Data_Jajarkot_Rukum_West/4209-d298c5a6-b7f8-4615-bcda-bf3b7d301810_raw_images",
+  "datasets/raw-images/4_Sample_Data_Jajarkot_Rukum_West/4210-64fe11ff-2dc5-42d2-aad8-a67463b17c9b_raw_images"
 ]
